requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sqlalchemy>=2.0.0
deepface>=0.0.89
numpy>=1.24.0
Pillow>=10.0.0
tf-keras
